reCAPTCHA v3 only
So this early alpha build has only the reCAPTCHA v3 check enabled, so when the "Solve reCAPTCHA" windows opens, wait till you see the recaptcha logo in the bottom right, then close the window. In the terminal you'll get a "reCAPTCHA v3 score" from 0.0 to 1.0. This value needs to be at least 0.5 or higher to be able to login!